## Enable Server Caching by Default with Improved 2-Layer Architecture

This PR significantly improves Weave's caching system by enabling server response caching by default and implementing a robust 2-layer cache architecture for better performance and reliability.

### 🚀 Key Changes

**Default Caching Enabled**
- Changed `WEAVE_USE_SERVER_CACHE` default from `false` to `true`
- Server responses now cached by default for improved performance

**Enhanced Write Endpoint Caching** 
- Added caching to `file_create` and `obj_create` endpoints
- These endpoints generate stable digests making them naturally idempotent and safe to cache
- Significantly improves write performance for repeated operations

**New 2-Layer Cache Architecture**
- **Memory Layer**: Fast LRU cache for process-lifetime caching
- **Disk Layer**: Persistent storage for cross-process/session caching  
- **Stacked Design**: Automatic read-through with faster layer population

**Performance Optimizations**
- **Existence Check Optimization**: Skip redundant writes when same key exists (leverages digest stability)
- **Silent Error Handling**: Robust error handling that degrades gracefully

**Test Infrastructure Improvements**
- **Cache Isolation**: New `caching_client_isolation` fixture prevents cross-test contamination
- **Comprehensive Coverage**: 8 new tests covering default behavior, persistence, error handling, and layer interaction
- **Robustness Testing**: Validates cache works correctly under various failure scenarios

**Code Quality Improvements**
- **Clean Architecture**: Protocol-based design with `LRUCache`, `DiskCache`, and `StackedCache` classes
- **Better Abstraction**: Moved from monolithic implementation to composable cache layers
- **Improved Maintainability**: Clear separation of concerns and better error handling

**Bug Fixes**
- Fixed `future_executor_fastlane` construction issue
- Resolved threading and WAL consistency issues from previous implementation

### 🎯 Benefits

- **Faster Performance**: Memory cache provides sub-millisecond access for hot data
- **Cross-Session Persistence**: Disk cache survives process restarts  
- **Better User Experience**: Caching enabled by default means users get performance benefits out-of-the-box
- **Robust Operation**: Silent error handling ensures cache failures don't break functionality
- **Easy Testing**: Cache isolation prevents test interference and flaky behavior

### 📊 Impact

This change makes caching a first-class feature in Weave, significantly improving performance for users working with repeated file/object operations while maintaining full backward compatibility.

Simple test:
```python
import weave
import random
import time
from weave import EvaluationLogger

client = weave.init("eval-log-perf")
eval_logger = EvaluationLogger(
    model="llama3_8b",
    dataset="openwebtext",
)

print("Logging 100 predictions")
start_time = time.time()
for i in range(100):
    pred_logger = eval_logger.log_prediction(
        inputs={
            "prompt": "Hello, " * random.randint(10, 1000),
        },
        output="World!" * random.randint(10, 100)
    )
    pred_logger.log_score(scorer="success", score=random.choice([True, False]))
    pred_logger.finish()

end_time = time.time()
print(f"Logging time: {end_time - start_time}s")

eval_logger.log_summary({
    "latency_ms": {
        "mean": 100,
        "std": 10,
    },
    "ttft_ms": {
        "mean": 100,
        "std": 10,
    },
})
client.finish()
finish_time = time.time()
print(f"Flushing time: {finish_time - end_time}s")
```

* `WEAVE_CLIENT_PARALLELISM=0 python perf.py` (2x speedup)
    * This PR: Logging time: 30.302637815475464s, Flushing time: 0.002299070358276367s
    * Master: Logging time: 61.64315414428711s, Flushing time: 2.0204007625579834s
* `WEAVE_CLIENT_PARALLELISM=2 python perf.py` (2x speedup)
    * This PR: Logging time: 0.12539410591125488s, Flushing time: 31.96068286895752s
    * Master: Logging time: 0.1964430809020996s, Flushing time: 63.38675594329834s
* `WEAVE_CLIENT_PARALLELISM=32 python perf.py` (5x speedup)
    * This PR: Logging time: 0.14585304260253906s, Flushing time: 2.5890707969665527s
    * Master: Logging time: 0.2528226375579834s, Flushing time: 12.966535091400146s
* `WEAVE_CLIENT_PARALLELISM=1000 python perf.py` (5x speedup)
    * This PR: Logging time: 0.3197591304779053s, Flushing time: 0.623661994934082s
    * Master: Logging time: 0.3289499282836914s, Flushing time: 3.1006669998168945s